changes accordingly comments
